+++ cmd/deploy/deploy_test.go
@@ -2,14 +2,20 @@ package deploy
 
 import (
 	"context"
+	"io"
+	"os"
+	"strings"
 	"testing"
+	"time"
 
+	"numerous.com/cli/cmd/output"
 	"numerous.com/cli/internal/app"
 	"numerous.com/cli/internal/appident"
 	"numerous.com/cli/internal/test"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
 )
 
 func TestDeploy(t *testing.T) {
@@ -20,20 +26,28 @@ func TestDeploy(t *testing.T) {
 	const uploadURL = "https://upload/url"
 	const deployVersionID = "deploy-version-id"
 
-	mockVersionDeploy := func(apps *mockAppService) {
+	mockVersionDeployWithDeployEventsRun := func(apps *mockAppService, deployEventsRun func(mock.Arguments)) {
 		apps.On("CreateVersion", mock.Anything, mock.Anything).Return(app.CreateAppVersionOutput{AppVersionID: appVersionID}, nil)
 		apps.On("AppVersionUploadURL", mock.Anything, mock.Anything).Return(app.AppVersionUploadURLOutput{UploadURL: uploadURL}, nil)
 		apps.On("UploadAppSource", mock.Anything, mock.Anything).Return(nil)
 		apps.On("DeployApp", mock.Anything, mock.Anything).Return(app.DeployAppOutput{DeploymentVersionID: deployVersionID}, nil)
-		apps.On("DeployEvents", mock.Anything, mock.Anything).Return(nil)
+		apps.On("DeployEvents", mock.Anything, mock.Anything).Run(deployEventsRun).Return(nil)
+	}
+
+	mockVersionDeploy := func(apps *mockAppService) {
+		mockVersionDeployWithDeployEventsRun(apps, nil)
+	}
+
+	mockAppExistsWithDeployEventsRun := func(deployEventsRun func(mock.Arguments)) *mockAppService {
+		apps := &mockAppService{}
+		apps.On("ReadApp", mock.Anything, mock.Anything).Return(app.ReadAppOutput{AppID: appID}, nil)
+		mockVersionDeployWithDeployEventsRun(apps, deployEventsRun)
+
+		return apps
 	}
 
 	mockAppExists := func() *mockAppService {
-		apps := &mockAppService{}
-		apps.On("ReadApp", mock.Anything, mock.Anything).Return(app.ReadAppOutput{AppID: appID}, nil)
-		mockVersionDeploy(apps)
-
-		return apps
+		return mockAppExistsWithDeployEventsRun(nil)
 	}
 
 	mockAppNotExists := func() *mockAppService {
@@ -196,4 +210,127 @@ func TestDeploy(t *testing.T) {
 			apps.AssertCalled(t, "CreateVersion", mock.Anything, expectedInput)
 		}
 	})
+
+	t.Run("prints expected verbose messages", func(t *testing.T) {
+		appDir := t.TempDir()
+		test.CopyDir(t, "../../testdata/streamlit_app", appDir)
+		expectedVersion := "v1.2.3"
+		expectedMessage := "expected message"
+		apps := mockAppExistsWithDeployEventsRun(func(args mock.Arguments) {
+			input := args.Get(1).(app.DeployEventsInput)
+			input.Handler(app.DeployEvent{Typename: "AppBuildMessageEvent", BuildMessage: app.AppBuildMessageEvent{Message: "Build message 1"}})    // nolint:errcheck
+			input.Handler(app.DeployEvent{Typename: "AppBuildMessageEvent", BuildMessage: app.AppBuildMessageEvent{Message: "Build message 2"}})    // nolint:errcheck
+			input.Handler(app.DeployEvent{Typename: "AppDeploymentStatusEvent", DeploymentStatus: app.AppDeploymentStatusEvent{Status: "PENDING"}}) // nolint:errcheck
+			input.Handler(app.DeployEvent{Typename: "AppDeploymentStatusEvent", DeploymentStatus: app.AppDeploymentStatusEvent{Status: "PENDING"}}) // nolint:errcheck
+			input.Handler(app.DeployEvent{Typename: "AppDeploymentStatusEvent", DeploymentStatus: app.AppDeploymentStatusEvent{Status: "PENDING"}}) // nolint:errcheck
+			input.Handler(app.DeployEvent{Typename: "AppDeploymentStatusEvent", DeploymentStatus: app.AppDeploymentStatusEvent{Status: "RUNNING"}}) // nolint:errcheck
+		})
+
+		input := DeployInput{AppDir: appDir, OrgSlug: slug, AppSlug: appSlug, Version: expectedVersion, Message: expectedMessage, Verbose: true}
+
+		stdout, err := mockStdout(t, func() error {
+			return Deploy(context.TODO(), apps, input)
+		})
+
+		assert.NoError(t, err)
+		expected := []string{
+			"<non-ascii> Loading app configuration............................",
+			"\r<non-ascii> Loading app configuration............................OK\n",
+			"<non-ascii> Registering new version for organization-slug/app-...",
+			"\r<non-ascii> Registering new version for organization-slug/app-...OK\n",
+			"<non-ascii> Creating app archive.................................",
+			"\r<non-ascii> Creating app archive.................................OK\n",
+			"<non-ascii> Uploading app archive................................",
+			"\r<non-ascii> Uploading app archive................................OK\n",
+			"<non-ascii> Deploying app........................................\n",
+			"Build Build message 1\n",
+			"Build Build message 2\n",
+			"\rDeploy Workload is pending",
+			"\rDeploy Workload is pending.",
+			"\rDeploy Workload is pending..\n",
+			"\rDeploy Workload is running\n",
+			"<non-ascii> Deploying app........................................OK\n",
+			"<non-ascii> Access your app at: https://www.numerous.com/app/organization/organization-slug/private/app-slug\n",
+			"\n",
+			"To read the logs from your app you can:\n",
+			"  numerous logs --organization=organization-slug --app=app-slug\n",
+			"Or you can use the --follow flag:\n",
+			"  numerous deploy --follow --organization=organization-slug --app=app-slug " + appDir + "\n",
+		}
+		output, _ := io.ReadAll(stdout)
+		actual := cleanNonASCIIAndANSI(string(output))
+		assert.Equal(t, strings.Join(expected, ""), actual)
+	})
+
+	t.Run("given follow flag it reads deployment logs", func(t *testing.T) {
+		appDir := t.TempDir()
+		test.CopyDir(t, "../../testdata/streamlit_app", appDir)
+		expectedVersion := "v1.2.3"
+		expectedMessage := "expected message"
+		apps := mockAppExists()
+		ch := make(chan app.AppDeployLogEntry, 2)
+		expectedEntry1 := app.AppDeployLogEntry{Timestamp: time.Date(2024, time.January, 1, 1, 1, 1, 0, time.UTC), Text: "Log entry 1"}
+		expectedEntry2 := app.AppDeployLogEntry{Timestamp: time.Date(2024, time.January, 1, 1, 1, 2, 0, time.UTC), Text: "Log entry 2"}
+		ch <- expectedEntry1
+		ch <- expectedEntry2
+		close(ch)
+		apps.On("AppDeployLogs", appident.AppIdentifier{OrganizationSlug: slug, AppSlug: appSlug}).Once().Return(ch, nil)
+
+		input := DeployInput{AppDir: appDir, OrgSlug: slug, AppSlug: appSlug, Version: expectedVersion, Message: expectedMessage, Verbose: true, Follow: true}
+		stdout, err := mockStdout(t, func() error {
+			return Deploy(context.TODO(), apps, input)
+		})
+
+		assert.NoError(t, err)
+		output, _ := io.ReadAll(stdout)
+		actual := cleanNonASCIIAndANSI(string(output))
+
+		expected := strings.Join(
+			[]string{
+				"Following logs of organization-slug/app-slug:",
+				expectedEntry1.Timestamp.Format(time.RFC3339Nano) + " " + expectedEntry1.Text,
+				expectedEntry2.Timestamp.Format(time.RFC3339Nano) + " " + expectedEntry2.Text,
+			}, "\n",
+		)
+		assert.Contains(t, actual, expected)
+		apps.AssertExpectations(t)
+	})
+}
+
+// Strips output of known ANSI terminal escapes, and non-ascii runes (e.g. icons).
+func cleanNonASCIIAndANSI(s string) string {
+	var cleaned string
+	for _, r := range s {
+		if r < 128 {
+			cleaned += string(r)
+		} else {
+			cleaned += "<non-ascii>"
+		}
+	}
+
+	for _, code := range []string{output.AnsiRed, output.AnsiReset, output.AnsiGreen, output.AnsiFaint, output.AnsiCyanBold} {
+		cleaned = strings.ReplaceAll(cleaned, code, "")
+	}
+
+	return cleaned
+}
+
+func mockStdout(t *testing.T, f func() error) (io.Reader, error) {
+	t.Helper()
+
+	realStdout := os.Stdout
+
+	defer func() {
+		os.Stdout = realStdout
+	}()
+
+	r, w, err := os.Pipe()
+	require.NoError(t, err)
+	os.Stdout = w
+
+	err = f()
+
+	require.NoError(t, w.Close())
+
+	return r, err
 }