+++ internal/app/upload_app_source.go
@@ -2,12 +2,21 @@ package app
 
 import (
 	"bytes"
-	"errors"
+	"fmt"
 	"io"
 	"net/http"
 )
 
-var ErrAppSourceUpload = errors.New("error uploading app source")
+type AppSourceUploadError struct {
+	HTTPStatusCode int
+	HTTPStatus     string
+	UploadURL      string
+	ResponseBody   []byte
+}
+
+func (e *AppSourceUploadError) Error() string {
+	return fmt.Sprintf("http %d: %q uploading app source file to %q ", e.HTTPStatusCode, e.HTTPStatus, e.UploadURL)
+}
 
 func (s *Service) UploadAppSource(uploadURL string, archive io.Reader) error {
 	var buf bytes.Buffer
@@ -26,7 +35,13 @@ func (s *Service) UploadAppSource(uploadURL string, archive io.Reader) error {
 	}
 
 	if resp.StatusCode != http.StatusOK {
-		return ErrAppSourceUpload
+		responseBody, _ := io.ReadAll(resp.Body)
+		return &AppSourceUploadError{
+			HTTPStatusCode: resp.StatusCode,
+			HTTPStatus:     resp.Status,
+			UploadURL:      uploadURL,
+			ResponseBody:   responseBody,
+		}
 	}
 
 	return nil